feat: Add SongEditType enum for song editing actions
The `SongEditType` enum is added to the `type.d.ts` file in the `lib/express` directory. This enum defines various actions that can be performed on a song, such as pause, resume, skip, stop, addSong, setTime, removeSong, setVolume, setQueue, quit, and seek. This addition enhances the express functionality of the Amhra application by providing a standardized way to handle song editing actions.