feat: Add isMuting check to setVolume method in CustomAudioPlayer
This commit adds an isMuting check to the setVolume method in the CustomAudioPlayer class. Previously, the volume would be set even when the player was muted, resulting in unintended behavior. With this change, the volume will only be set if the player is not muted. This improves the functionality and user experience of the audio player.