feat: Update start time in CustomAudioPlayer
This commit updates the `playResource` method in the `CustomAudioPlayer` class to correctly update the start time before playing a resource. Previously, the start time was set to the current time before checking if the resource had a specified start time. With this change, the start time is only updated if the resource has a start time, ensuring accurate playback.